renames project to p0
