+++ .gitignore
@@ -26,3 +26,7 @@ replay_pid*
 *.settings/
 *.classpath
 *.project
+
+*target/
+*.properties
+*DS_Store/
