+++ src/main/java/com/revature/p0/p0.java
@@ -1,6 +1,6 @@
-package com.revature.app;
+package com.revature.p0;
 
-public class App {
+public class p0 {
   public static void main(String args[]) {
     System.out.println("Welcome to eCommerce App!");
   }
