+++ src/test/java/com/revature/p0/AppTest.java
@@ -1,4 +1,4 @@
-package com.mycompany.app;
+package com.revature.p0;
 
 import static org.junit.Assert.assertTrue;
 
@@ -7,14 +7,12 @@ import org.junit.Test;
 /**
  * Unit test for simple App.
  */
-public class AppTest 
-{
+public class AppTest {
     /**
      * Rigorous Test :-)
      */
     @Test
-    public void shouldAnswerWithTrue()
-    {
-        assertTrue( true );
+    public void shouldAnswerWithTrue() {
+        assertTrue(true);
     }
 }
